Refactor for project being moved into a cluster ns
We made project a namespaced resource. Now, projects are created
under a cluster's namespace in the mangement plane. This cause's
the project's ID to be of the form `<cluster name>:<proejct name>`.
That change in format also necessitated moving the project ID
from a label to an annotation on namespaces because `:` is not a
valid character for label values.

To handle the new format, we needed to switch to using a custom
index instead of the labels-based cache.